refactor circular dependencies
