+++ src/databricks/labs/ucx/contexts/cli_command.py
@@ -17,6 +17,7 @@ from databricks.labs.ucx.azure.credentials import ServicePrincipalMigration, Sto
 from databricks.labs.ucx.azure.locations import ExternalLocationsMigration
 from databricks.labs.ucx.azure.resources import AzureAPIClient, AzureResources
 from databricks.labs.ucx.contexts.application import GlobalContext
+from databricks.labs.ucx.source_code.notebooks.loaders import NotebookLoader, LocalNotebookLoader
 from databricks.labs.ucx.source_code.files import LocalFileMigrator
 from databricks.labs.ucx.workspace_access.clusters import ClusterAccess
 
@@ -165,6 +166,10 @@ class WorkspaceContext(CliContext):
             self.aws_resource_permissions,
             self.iam_credential_manager,
         )
+
+    @cached_property
+    def notebook_loader(self) -> NotebookLoader:
+        return LocalNotebookLoader(self.syspath_provider)
 
 
 class AccountContext(CliContext):
