+++ src/databricks/labs/ucx/source_code/base.py
@@ -110,6 +110,3 @@ class SequentialLinter(Linter):
     def lint(self, code: str) -> Iterable[Advice]:
         for linter in self._linters:
             yield from linter.lint(code)
-
-
-NOTEBOOK_HEADER = "Databricks notebook source"
