+++ src/databricks/labs/ucx/source_code/notebooks/sources.py
@@ -1,0 +1,98 @@
+from __future__ import annotations
+
+from collections.abc import Iterable
+
+from databricks.sdk.service.workspace import Language
+
+from databricks.labs.ucx.hive_metastore.migration_status import MigrationIndex
+from databricks.labs.ucx.source_code.base import Advice
+
+from databricks.labs.ucx.source_code.graph import SourceContainer, DependencyGraph, DependencyProblem
+from databricks.labs.ucx.source_code.languages import Languages
+from databricks.labs.ucx.source_code.notebooks.cells import CellLanguage, Cell, NOTEBOOK_HEADER, CELL_SEPARATOR
+
+
+class Notebook(SourceContainer):
+
+    @staticmethod
+    def parse(path: str, source: str, default_language: Language) -> Notebook:
+        default_cell_language = CellLanguage.of_language(default_language)
+        cells = default_cell_language.extract_cells(source)
+        if cells is None:
+            raise ValueError(f"Could not parse Notebook: {path}")
+        return Notebook(path, source, default_language, cells, source.endswith('\n'))
+
+    def __init__(self, path: str, source: str, language: Language, cells: list[Cell], ends_with_lf):
+        self._path = path
+        self._source = source
+        self._language = language
+        self._cells = cells
+        self._ends_with_lf = ends_with_lf
+
+    @property
+    def path(self) -> str:
+        return self._path
+
+    @property
+    def cells(self) -> list[Cell]:
+        return self._cells
+
+    @property
+    def original_code(self) -> str:
+        return self._source
+
+    def to_migrated_code(self):
+        default_language = CellLanguage.of_language(self._language)
+        header = f"{default_language.comment_prefix} {NOTEBOOK_HEADER}"
+        sources = [header]
+        for i, cell in enumerate(self._cells):
+            migrated_code = cell.migrated_code
+            if cell.language is not default_language:
+                migrated_code = default_language.wrap_with_magic(migrated_code, cell.language)
+            sources.append(migrated_code)
+            if i < len(self._cells) - 1:
+                sources.append('')
+                sources.append(f'{default_language.comment_prefix} {CELL_SEPARATOR}')
+                sources.append('')
+        if self._ends_with_lf:
+            sources.append('')  # following join will append lf
+        return '\n'.join(sources)
+
+    def build_dependency_graph(self, parent: DependencyGraph) -> None:
+        problems: list[DependencyProblem] = []
+        for cell in self._cells:
+            cell.build_dependency_graph(parent, problems.append)
+        parent.add_problems(problems)
+
+
+class NotebookLinter:
+    """
+    Parses a Databricks notebook and then applies available linters
+    to the code cells according to the language of the cell.
+    """
+
+    def __init__(self, langs: Languages, notebook: Notebook):
+        self._languages: Languages = langs
+        self._notebook: Notebook = notebook
+
+    @classmethod
+    def from_source(cls, index: MigrationIndex, source: str, default_language: Language) -> 'NotebookLinter':
+        langs = Languages(index)
+        notebook = Notebook.parse("", source, default_language)
+        assert notebook is not None
+        return cls(langs, notebook)
+
+    def lint(self) -> Iterable[Advice]:
+        for cell in self._notebook.cells:
+            if not self._languages.is_supported(cell.language.language):
+                continue
+            linter = self._languages.linter(cell.language.language)
+            for advice in linter.lint(cell.original_code):
+                yield advice.replace(
+                    start_line=advice.start_line + cell.original_offset,
+                    end_line=advice.end_line + cell.original_offset,
+                )
+
+    @staticmethod
+    def name() -> str:
+        return "notebook-linter"
