+++ tests/unit/source_code/test_s3fs.py
@@ -2,19 +2,16 @@ from pathlib import Path
 
 import pytest
 
-from databricks.labs.ucx.source_code.dependency_graph import (
+from databricks.labs.ucx.source_code.graph import (
+    SourceContainer,
+    DependencyResolver,
+    DependencyProblem,
     DependencyGraphBuilder,
 )
-from databricks.labs.ucx.source_code.dependency_problem import DependencyProblem
-from databricks.labs.ucx.source_code.dependency_loaders import LocalFileLoader, LocalNotebookLoader
-from databricks.labs.ucx.source_code.dependency_containers import SourceContainer
-from databricks.labs.ucx.source_code.dependency_resolvers import DependencyResolver
-from databricks.labs.ucx.source_code.site_packages import SitePackages
-from databricks.labs.ucx.source_code.syspath_provider import SysPathProvider
-from databricks.labs.ucx.source_code.whitelist import Whitelist
+from databricks.labs.ucx.source_code.files import FileLoader, LocalFileResolver
+from databricks.labs.ucx.source_code.whitelist import WhitelistResolver, Whitelist
 from tests.unit import (
     _load_sources,
-    locate_site_packages,
     _local_loader_with_side_effects,
 )
 
@@ -113,20 +110,21 @@ def test_detect_s3fs_import(empty_index, source: str, expected: list[DependencyP
     datas = _load_sources(SourceContainer, "s3fs-python-compatibility-catalog.yml")
     whitelist = Whitelist.parse(datas[0])
     sources = {"path": source}
-    file_loader = _local_loader_with_side_effects(LocalFileLoader, sources, {})
-    site_packages = SitePackages.parse(locate_site_packages())
-    provider = SysPathProvider.from_pathlike_string("")
-    resolver = DependencyResolver.initialize(
-        whitelist, site_packages, file_loader, LocalNotebookLoader(provider), provider
+    file_loader = _local_loader_with_side_effects(FileLoader, sources, {})
+    dependency_resolver = DependencyResolver(
+        [
+            WhitelistResolver(whitelist),
+            LocalFileResolver(file_loader),
+        ]
     )
-    builder = DependencyGraphBuilder(resolver)
+    builder = DependencyGraphBuilder(dependency_resolver)
     builder.build_local_file_dependency_graph(Path("path"))
-    problems: list[DependencyProblem] = list(resolver.problems)
+    problems: list[DependencyProblem] = list(dependency_resolver.problems)
     assert problems == expected
 
 
 @pytest.mark.parametrize(
-    " expected",
+    "expected",
     (
         [
             DependencyProblem(
@@ -145,13 +143,14 @@ def test_detect_s3fs_import_in_dependencies(empty_index, expected: list[Dependen
     sources: dict[str, str] = dict(zip(paths, _load_sources(SourceContainer, *paths)))
     datas = _load_sources(SourceContainer, "s3fs-python-compatibility-catalog.yml")
     whitelist = Whitelist.parse(datas[0])
-    file_loader = _local_loader_with_side_effects(LocalFileLoader, sources, {})
-    site_packages = SitePackages.parse(locate_site_packages())
-    provider = SysPathProvider.from_pathlike_string("")
-    resolver = DependencyResolver.initialize(
-        whitelist, site_packages, file_loader, LocalNotebookLoader(provider), provider
+    file_loader = _local_loader_with_side_effects(FileLoader, sources, {})
+    dependency_resolver = DependencyResolver(
+        [
+            WhitelistResolver(whitelist),
+            LocalFileResolver(file_loader),
+        ]
     )
-    builder = DependencyGraphBuilder(resolver)
+    builder = DependencyGraphBuilder(dependency_resolver)
     builder.build_local_file_dependency_graph(Path("root9.py.txt"))
-    problems: list[DependencyProblem] = list(resolver.problems)
+    problems: list[DependencyProblem] = list(dependency_resolver.problems)
     assert problems == expected
